refactor: separate test application screens
